fix setting of delivery number

+++ qpid-proton4j-engine/src/main/java/org/apache/qpid/proton4j/engine/impl/ProtonOutgoingDelivery.java
@@ -192,7 +192,7 @@ public class ProtonOutgoingDelivery implements OutgoingDelivery {
         return deliveryId;
     }
 
-    void setDeliveryId(byte deliveryId) {
+    void setDeliveryId(long deliveryId) {
         this.deliveryId = deliveryId;
     }
